Menu has been added.
After login of the users, they are moved to the waiting room, in which they can explore the leaderboard, settings, and other information.
In the menu there is the PLAY option, if selected, users will be moved to the VirtualViewAssignment and a new match will be instantiated.

Disconnections are handled both in case of client program closed or in case of network communication issues. Automatic reconnection is performed.
Waiting for loading of matched from file to complete the implementation.