Added mute buttons in PlayingPane, moved the help request into the client and modified the SoundManager.
Added a top bar in the playingPane where were placed two mute buttons, one for the effects and one for the music. The help action was wrongly in the server side and now it is moved in to the client. Added static variables to contain the path of the songs and sound effects in the SoundManager.